Change documentation wordpress-server-vhosts.conf
Does the file "wordpress-server-vhosts.conf" was mistyped?
I think it mean to be "wordpress-server-block.conf", because this environment is Nginx based.

I was trying to follow instructions on documentation, to change my current environment that was Apache based to use this image instead.
But I saw several times "wordpress-server-vhosts.conf" instead of "wordpress-server-block.conf" on docker-compose.yml examples.
Is that right? Or we really need to update the documentation?